# RyuJIT: Optimize "X<<1" to "lea [X+X]"