A small peephole optimization `X<<1` to `lea [X+X]`: (**UPD** or `add X, X`)
```csharp
static int Test1(int x) => x * 2; // optimized to "x<<1"
static int Test2(int x) => x << 1;
```
#### Current codegen:
```asm
; Method Test1(int):int
       mov      eax, ecx
       shl      eax, 1
       ret      
; Total bytes of code: 5


; Method Test2(int):int
       mov      eax, ecx
       shl      eax, 1
       ret      
; Total bytes of code: 5
```
#### New codegen:
```asm
; Method Test1(int):int
       lea      eax, [rcx+rcx]
       ret      
; Total bytes of code: 4


; Method Test2(int):int
       lea      eax, [rcx+rcx]
       ret      
; Total bytes of code: 4
```
#### Jit-diff (-f -pmi):
```
Found 52 files with textual diffs.

Summary of Code Size diffs:
(Lower is better)

Total bytes of diff: -654 (-0.00% of base)
    diff is an improvement.

Top file improvements (bytes):
        -184 : System.Private.CoreLib.dasm (-0.00% of base)
        -116 : Microsoft.Diagnostics.Tracing.TraceEvent.dasm (-0.00% of base)
         -70 : System.Private.Xml.dasm (-0.00% of base)
         -33 : System.Private.DataContractSerialization.dasm (-0.00% of base)
         -30 : System.Collections.dasm (-0.01% of base)
         -24 : System.Private.Xml.Linq.dasm (-0.02% of base)
         -15 : System.Data.OleDb.dasm (-0.00% of base)
         -15 : System.Linq.dasm (-0.00% of base)
         -15 : System.Private.Uri.dasm (-0.02% of base)
         -14 : System.Linq.Parallel.dasm (-0.00% of base)
         -10 : System.Data.Odbc.dasm (-0.01% of base)
         -10 : System.Text.RegularExpressions.dasm (-0.00% of base)
          -8 : Microsoft.CodeAnalysis.CSharp.dasm (-0.00% of base)
          -8 : Microsoft.CodeAnalysis.dasm (-0.00% of base)
          -8 : Newtonsoft.Json.dasm (-0.00% of base)
          -6 : Microsoft.CodeAnalysis.VisualBasic.dasm (-0.00% of base)
          -6 : System.Drawing.Common.dasm (-0.00% of base)
          -6 : System.IO.FileSystem.dasm (-0.00% of base)
          -6 : System.Runtime.Numerics.dasm (-0.01% of base)
          -5 : System.Reflection.Metadata.dasm (-0.00% of base)

51 total files with Code Size differences (51 improved, 0 regressed), 109 unchanged.

Top method improvements (bytes):
         -10 (-2.48% of base) : System.Private.CoreLib.dasm - System.String:Insert(int,System.String):System.String:this
          -8 (-0.27% of base) : System.Private.CoreLib.dasm - System.String:ReplaceHelper(int,System.String,System.ReadOnlySpan`1[Int32]):System.String:this (2 methods)
          -8 (-1.16% of base) : System.Private.Uri.dasm - System.Uri:Compress(System.Char[],int,byref,System.UriParser):System.Char[]
          -6 (-2.62% of base) : System.Private.CoreLib.dasm - System.Span`1[Int16][System.Int16]:Fill(short):this
          -6 (-1.57% of base) : System.Private.DataContractSerialization.dasm - System.Xml.XmlBinaryNodeWriter:UnsafeWriteText(long,int):this
          -6 (-0.63% of base) : System.Private.Xml.dasm - System.Xml.XmlTextReaderImpl:ParsePIValue(byref,byref):bool:this
          -6 (-0.34% of base) : System.Private.Xml.dasm - System.Xml.XmlTextReaderImpl:ParseText(byref,byref,byref):bool:this
          -6 (-0.41% of base) : System.Private.Xml.dasm - <ParsePIValueAsync>d__562:MoveNext():this
          -6 (-0.37% of base) : System.Private.Xml.dasm - <ParseCDataOrCommentTupleAsync>d__566:MoveNext():this
          -5 (-1.78% of base) : System.Private.CoreLib.dasm - System.String:PadRight(int,ushort):System.String:this
          -5 (-0.91% of base) : System.Private.CoreLib.dasm - System.String:Remove(int,int):System.String:this
          -5 (-0.33% of base) : System.Private.CoreLib.dasm - System.String:JoinCore(long,int,System.String[],int,int):System.String
          -4 (-0.85% of base) : Newtonsoft.Json.dasm - Newtonsoft.Json.JsonTextReader:PrepareBufferForReadData(bool,int):this
          -4 (-0.16% of base) : System.IO.FileSystem.dasm - System.IO.Enumeration.FileSystemName:MatchPattern(System.ReadOnlySpan`1[Char],System.ReadOnlySpan`1[Char],bool,bool):bool
          -4 (-0.05% of base) : System.Numerics.Tensors.dasm - System.Numerics.Tensors.CompressedSparseTensor`1[Int16][System.Int16]:EnsureCapacity(int,int):this
          -4 (-0.12% of base) : System.Private.CoreLib.dasm - System.Number:Dragon4(long,int,int,bool,int,bool,System.Span`1[Byte],byref):int
          -4 (-5.19% of base) : System.Private.Xml.dasm - System.Xml.IncrementalReadCharsDecoder:Decode(System.Char[],int,int):int:this
          -4 (-0.37% of base) : System.Runtime.Numerics.dasm - System.Numerics.BigIntegerCalculator:Square(long,int,long,int)
          -3 (-0.16% of base) : Microsoft.CodeAnalysis.dasm - Microsoft.Cci.MetadataWriter:PopulateMethodSemanticsTableRows():this
          -3 (-0.07% of base) : Microsoft.Diagnostics.Tracing.TraceEvent.dasm - Microsoft.Diagnostics.Tracing.Session.TraceEventSession:EnableProvider(System.Guid,int,long,Microsoft.Diagnostics.Tracing.Session.TraceEventProviderOptions):bool:this

Top method improvements (percentages):
          -1 (-20.00% of base) : System.Reflection.Metadata.dasm - System.Reflection.Metadata.Ecma335.HasSemanticsTag:ConvertEventHandleToTag(System.Reflection.Metadata.EventDefinitionHandle):int
          -1 (-20.00% of base) : System.Reflection.Metadata.dasm - System.Reflection.Metadata.Ecma335.TypeOrMethodDefTag:ConvertTypeDefRowIdToTag(System.Reflection.Metadata.TypeDefinitionHandle):int
          -1 (-12.50% of base) : System.Reflection.Metadata.dasm - System.Reflection.Metadata.Ecma335.HasSemanticsTag:ConvertPropertyHandleToTag(System.Reflection.Metadata.PropertyDefinitionHandle):int
          -1 (-12.50% of base) : System.Reflection.Metadata.dasm - System.Reflection.Metadata.Ecma335.MemberForwardedTag:ConvertMethodDefToTag(System.Reflection.Metadata.MethodDefinitionHandle):int
          -1 (-12.50% of base) : System.Reflection.Metadata.dasm - System.Reflection.Metadata.Ecma335.TypeOrMethodDefTag:ConvertMethodDefToTag(System.Reflection.Metadata.MethodDefinitionHandle):int
          -1 (-7.69% of base) : Microsoft.Diagnostics.Tracing.TraceEvent.dasm - Microsoft.Diagnostics.Tracing.Analysis.GC.GCCondemnedReasons:GetReasonWithGenNumber(int):int:this
          -1 (-6.67% of base) : System.ComponentModel.TypeConverter.dasm - System.ComponentModel.InterlockedBitVector32:CreateMask(int):int
          -4 (-5.19% of base) : System.Private.Xml.dasm - System.Xml.IncrementalReadCharsDecoder:Decode(System.Char[],int,int):int:this
          -2 (-4.35% of base) : System.Linq.Parallel.dasm - System.Linq.Parallel.SortHelper`2[__Canon,Int64][System.__Canon,System.Int64]:ComputePartnerIndex(int):int:this
          -2 (-4.35% of base) : System.Linq.Parallel.dasm - System.Linq.Parallel.SortHelper`2[Byte,Int64][System.Byte,System.Int64]:ComputePartnerIndex(int):int:this
          -2 (-4.35% of base) : System.Linq.Parallel.dasm - System.Linq.Parallel.SortHelper`2[Int16,Int64][System.Int16,System.Int64]:ComputePartnerIndex(int):int:this
          -2 (-4.35% of base) : System.Linq.Parallel.dasm - System.Linq.Parallel.SortHelper`2[Int32,Int64][System.Int32,System.Int64]:ComputePartnerIndex(int):int:this
          -2 (-4.35% of base) : System.Linq.Parallel.dasm - System.Linq.Parallel.SortHelper`2[Double,Int64][System.Double,System.Int64]:ComputePartnerIndex(int):int:this
          -2 (-4.35% of base) : System.Linq.Parallel.dasm - System.Linq.Parallel.SortHelper`2[Vector`1,Int64][System.Numerics.Vector`1[System.Single],System.Int64]:ComputePartnerIndex(int):int:this
          -2 (-4.35% of base) : System.Linq.Parallel.dasm - System.Linq.Parallel.SortHelper`2[Int64,Int64][System.Int64,System.Int64]:ComputePartnerIndex(int):int:this
          -2 (-3.70% of base) : System.Private.Xml.dasm - System.Xml.XmlTextReaderImpl:ShiftBuffer(int,int,int):this
          -1 (-3.03% of base) : Microsoft.Diagnostics.Tracing.TraceEvent.dasm - Microsoft.Diagnostics.Tracing.Parsers.MicrosoftWindowsTCPIP.Ndkpi_BindArgs:get_NdkMr():long:this
          -1 (-3.03% of base) : Microsoft.Diagnostics.Tracing.TraceEvent.dasm - Microsoft.Diagnostics.Tracing.Parsers.MicrosoftWindowsTCPIP.Ndkpi_Complete_ConnectArgs:get_RequestContext():long:this
          -1 (-3.03% of base) : Microsoft.Diagnostics.Tracing.TraceEvent.dasm - Microsoft.Diagnostics.Tracing.Parsers.MicrosoftWindowsTCPIP.Ndkpi_Create_ListenerArgs:get_RequestContext():long:this
          -1 (-3.03% of base) : Microsoft.Diagnostics.Tracing.TraceEvent.dasm - Microsoft.Diagnostics.Tracing.Parsers.MicrosoftWindowsTCPIP.Ndkpi_Create_QpArgs:get_InitiatorCq():long:this

451 total methods with Code Size differences (451 improved, 0 regressed), 305951 unchanged.
```
I also tried:
`x << 2` -> `lea eax, [rcx*4]`
`x << 3` -> `lea eax, [rcx*8]`
but it was a size regression.