@EgorBo 2020-04-20
> Does this mean that it will generate `add eax, eax` if the destination register is the same as the source register?

Nope, it falls back to `shl eax, 1`
is `add eax, eax` better?