@GrabYourPitchforks 2020-04-21
According to Agner's tables, `add r,r` has better throughput characteristics and can be dispatched to a wider range of ports than `shl r,i`. But there might be subtleties I'm missing.