@BruceForstall 2020-04-22
Would a microbenchmark be able to measure the performance difference?