@EgorBo 2020-04-22
> Would a microbenchmark be able to measure the performance difference?

@BruceForstall I couldn't properly measure `lea` but `add reg, reg` seems shows some improvements:
```csharp
public class ShiftBenchmarks
{
    private int[] _array = Enumerable.Range(0, 10000).ToArray();

    [Benchmark]
    public void Benchmark_add()
    {
        int[] array = _array;
        for (int i = 0; i < array.Length; i++)
            array[i] = array[i] * 2; // "shl r9d, 1" vs "add r9d, r9d"
    }
}
```
```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18362.778 (1903/May2019Update/19H1)
Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores

       |        Method |     Mean |     Error |    StdDev | Code Size |
       |-------------- |---------:|----------:|----------:|----------:|
before | Benchmark_add | 4.302 us | 0.0139 us | 0.0123 us |      36 B |
after  | Benchmark_add | 3.350 us | 0.0110 us | 0.0097 us |      36 B |
```